requirement.txt
streamlit
pypdf
pdf2image
easyocr
opencv-python
pillow
sentence-transformers
faiss-cpu
ollama
python -m rag.vector_store
